Moving pear directory to /pear2/ for the moment.
It is more convenient to have this data outside of any directory that
contains source code that is being edited. If Bundles end up being
inside the vendor/ directory keeping pyrus in vendor/pyrus is no longer
an option since pyrus would back itself up prior to an installation and
thus never mark any packages as installed. This is caused by the way its
atomic file operations work. Unfortunately the php files cannot be stored
anywhere other than /pear2/php/ then, so an addition to the autoloader is
necessary to use PEAR2 dependencies.